// Frontend/Finlease-web/src/app/models/lease-request.model.ts
export class LeaseRequest{
    createdTime: Date;
    customerNumber: string;
    customerAccountNumber: string;
    customerName: string;
    phoneNumber: string;
    assetMake: string;
    assetModel: string;
    assetDeliveryDate: Date;
    assetPurpose: string;
    monthlyIncome: string;
    otherIncome: string;
    totalIncome: string;
    totalCommitmentMCB: string;
    totalCommitmentOther: string;
    creditTurnover: string;
    debitTurnover: string;
    setUpParameter: string;
    commentRecommendation: string;
    leaseType: string;
    requiredAmount: string;
    leasePeriodInMonth: string;
    createdBy: string;
    proposerUserId: string;
}